+++ cypress/integration/v2/blocks/blog-posts/desktop-style.spec.js
@@ -2,5 +2,10 @@
  * External dependencies
  */
 import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
 import { desktopStyle } from './'
+
 describe( 'Blog Posts Block ( Desktop style )', registerTests( [ desktopStyle ] ) )
