+++ cypress/integration/v2/blocks/blog-posts/mobile-style.spec.js
@@ -2,5 +2,9 @@
  * External dependencies
  */
 import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
 import { mobileStyle } from './'
 describe( 'Blog Posts Block ( Mobile style )', registerTests( [ mobileStyle ] ) )
