+++ cypress/integration/v2/blocks/blog-posts/tablet-advanced.spec.js
@@ -2,5 +2,9 @@
  * External dependencies
  */
 import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
 import { tabletAdvanced } from './'
 describe( 'Blog Posts Block ( Tablet advanced )', registerTests( [ tabletAdvanced ] ) )
