+++ cypress/integration/v2/blocks/blog-posts/tablet-style.spec.js
@@ -2,5 +2,10 @@
  * External dependencies
  */
 import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
 import { tabletStyle } from './'
+
 describe( 'Blog Posts Block ( Tablet style )', registerTests( [ tabletStyle ] ) )
